Del opp og oppdater casetester
