+++ src/test/kotlin/no/nav/familie/ba/sak/behandling/resultat/BehandlingsresultatUtilsTest.kt
@@ -434,192 +434,6 @@ class BehandlingsresultatUtilsTest {
         assertEquals(BehandlingResultat.AVSLÅTT_ENDRET_OG_OPPHØRT, behandlingsresultat)
     }
 
-
-    // Endringer uten søknad
-    @Test
-    fun `Scenarie 1, endret`() {
-        val behandlingsresultat = BehandlingsresultatUtils.utledBehandlingsresultatBasertPåYtelsePersoner(
-                listOf(
-                        YtelsePerson(
-                                personIdent = barn2Ident,
-                                ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
-                                kravOpprinnelse = KravOpprinnelse.TIDLIGERE,
-                                resultater = setOf(YtelsePersonResultat.ENDRET),
-                                ytelseSlutt = defaultYtelseSluttForLøpende,
-                        ),
-                        YtelsePerson(
-                                personIdent = barn1Ident,
-                                ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
-                                kravOpprinnelse = KravOpprinnelse.TIDLIGERE,
-                                resultater = setOf(YtelsePersonResultat.ENDRET),
-                                ytelseSlutt = defaultYtelseSluttForLøpende,
-                        )
-                )
-        )
-        assertEquals(BehandlingResultat.ENDRET, behandlingsresultat)
-    }
-
-    @Test
-    fun `Scenarie 2, endret og opphørt`() {
-        val behandlingsresultat = BehandlingsresultatUtils.utledBehandlingsresultatBasertPåYtelsePersoner(
-                listOf(
-                        YtelsePerson(
-                                personIdent = barn2Ident,
-                                ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
-                                kravOpprinnelse = KravOpprinnelse.TIDLIGERE,
-                                resultater = setOf(YtelsePersonResultat.ENDRET, YtelsePersonResultat.OPPHØRT),
-                                ytelseSlutt = inneværendeMåned()
-                        ),
-                        YtelsePerson(
-                                personIdent = barn1Ident,
-                                ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
-                                kravOpprinnelse = KravOpprinnelse.TIDLIGERE,
-                                resultater = setOf(YtelsePersonResultat.ENDRET, YtelsePersonResultat.OPPHØRT),
-                                ytelseSlutt = inneværendeMåned()
-                        )
-                )
-        )
-        assertEquals(BehandlingResultat.ENDRET_OG_OPPHØRT, behandlingsresultat)
-    }
-
-    @Test
-    fun `Scenarie 4, endret`() {
-        val behandlingsresultat = BehandlingsresultatUtils.utledBehandlingsresultatBasertPåYtelsePersoner(
-                listOf(
-                        YtelsePerson(
-                                personIdent = barn2Ident,
-                                ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
-                                kravOpprinnelse = KravOpprinnelse.TIDLIGERE,
-                                resultater = setOf(YtelsePersonResultat.ENDRET, YtelsePersonResultat.OPPHØRT),
-                                ytelseSlutt = inneværendeMåned()
-                        ),
-                        YtelsePerson(
-                                personIdent = barn1Ident,
-                                ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
-                                kravOpprinnelse = KravOpprinnelse.TIDLIGERE,
-                                resultater = setOf(),
-                                ytelseSlutt = defaultYtelseSluttForLøpende,
-                        )
-                )
-        )
-        assertEquals(BehandlingResultat.ENDRET, behandlingsresultat)
-    }
-
-    @Test
-    fun `Scenarie 5, endret`() {
-        val behandlingsresultat = BehandlingsresultatUtils.utledBehandlingsresultatBasertPåYtelsePersoner(
-                listOf(
-                        YtelsePerson(
-                                personIdent = barn2Ident,
-                                ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
-                                kravOpprinnelse = KravOpprinnelse.TIDLIGERE,
-                                resultater = setOf(YtelsePersonResultat.ENDRET),
-                                ytelseSlutt = defaultYtelseSluttForLøpende,
-                        ),
-                        YtelsePerson(
-                                personIdent = barn1Ident,
-                                ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
-                                kravOpprinnelse = KravOpprinnelse.TIDLIGERE,
-                                resultater = setOf(YtelsePersonResultat.ENDRET),
-                                ytelseSlutt = defaultYtelseSluttForLøpende,
-                        )
-                )
-        )
-        assertEquals(BehandlingResultat.ENDRET, behandlingsresultat)
-    }
-
-    @Test
-    fun `Scenarie 6, opphørt`() {
-        val behandlingsresultat = BehandlingsresultatUtils.utledBehandlingsresultatBasertPåYtelsePersoner(
-                listOf(
-                        YtelsePerson(
-                                personIdent = barn2Ident,
-                                ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
-                                kravOpprinnelse = KravOpprinnelse.TIDLIGERE,
-                                resultater = setOf(YtelsePersonResultat.OPPHØRT),
-                                ytelseSlutt = inneværendeMåned()
-                        ),
-                        YtelsePerson(
-                                personIdent = barn1Ident,
-                                ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
-                                kravOpprinnelse = KravOpprinnelse.TIDLIGERE,
-                                resultater = setOf(YtelsePersonResultat.OPPHØRT),
-                                ytelseSlutt = inneværendeMåned()
-                        )
-                )
-        )
-        assertEquals(BehandlingResultat.OPPHØRT, behandlingsresultat)
-    }
-
-    @Test
-    fun `Scenarie 8, endret og opphørt`() {
-        val behandlingsresultat = BehandlingsresultatUtils.utledBehandlingsresultatBasertPåYtelsePersoner(
-                listOf(
-                        YtelsePerson(
-                                personIdent = barn2Ident,
-                                ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
-                                kravOpprinnelse = KravOpprinnelse.TIDLIGERE,
-                                resultater = setOf(YtelsePersonResultat.ENDRET, YtelsePersonResultat.OPPHØRT),
-                                ytelseSlutt = inneværendeMåned()
-                        ),
-                        YtelsePerson(
-                                personIdent = barn1Ident,
-                                ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
-                                kravOpprinnelse = KravOpprinnelse.TIDLIGERE,
-                                resultater = setOf(YtelsePersonResultat.ENDRET, YtelsePersonResultat.OPPHØRT),
-                                ytelseSlutt = inneværendeMåned()
-                        )
-                )
-        )
-        assertEquals(BehandlingResultat.ENDRET_OG_OPPHØRT, behandlingsresultat)
-    }
-
-    @Test
-    fun `Scenarie 9, endret og opphørt`() {
-        val behandlingsresultat = BehandlingsresultatUtils.utledBehandlingsresultatBasertPåYtelsePersoner(
-                listOf(
-                        YtelsePerson(
-                                personIdent = barn2Ident,
-                                ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
-                                kravOpprinnelse = KravOpprinnelse.TIDLIGERE,
-                                resultater = setOf(YtelsePersonResultat.OPPHØRT),
-                                ytelseSlutt = inneværendeMåned().minusMonths(9)
-                        ),
-                        YtelsePerson(
-                                personIdent = barn1Ident,
-                                ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
-                                kravOpprinnelse = KravOpprinnelse.TIDLIGERE,
-                                resultater = setOf(YtelsePersonResultat.OPPHØRT),
-                                ytelseSlutt = inneværendeMåned()
-                        )
-                )
-        )
-        assertEquals(BehandlingResultat.ENDRET_OG_OPPHØRT, behandlingsresultat)
-    }
-
-    @Test
-    fun `Scenarie 10, endret`() {
-        val behandlingsresultat = BehandlingsresultatUtils.utledBehandlingsresultatBasertPåYtelsePersoner(
-                listOf(
-                        YtelsePerson(
-                                personIdent = barn2Ident,
-                                ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
-                                kravOpprinnelse = KravOpprinnelse.TIDLIGERE,
-                                resultater = setOf(YtelsePersonResultat.ENDRET),
-                                ytelseSlutt = defaultYtelseSluttForLøpende,
-                        ),
-                        YtelsePerson(
-                                personIdent = barn1Ident,
-                                ytelseType = YtelseType.ORDINÆR_BARNETRYGD,
-                                kravOpprinnelse = KravOpprinnelse.TIDLIGERE,
-                                resultater = setOf(),
-                                ytelseSlutt = defaultYtelseSluttForLøpende,
-                        )
-                )
-        )
-        assertEquals(BehandlingResultat.ENDRET, behandlingsresultat)
-    }
-
     @Test
     fun `Skal utlede opphørt når 2 barn blir redusert til samme måned`() {
         val behandlingsresultat = BehandlingsresultatUtils.utledBehandlingsresultatBasertPåYtelsePersoner(
